Added ROS service skills
Added an auto generator for skills that call a service. The result is stored
in the vars table and can be used in the exit method of the state. Utilities
have been added to innitialize service dependencies, execute a service in a
state and to create the skill module. Start script updated.