mongolog: create one document per skill string
Instead of having many entries for a single skill string interpretation we
now create a single document and then update and extend it while the skill
string is running. This allows for easier querying later, because one
document provides one full execution trace and timestamp.
Addtionally we now store the graph and skill status values. The graph is
recorded only once when the execution has ended for performance reasons.
The API for active skills has been changed again slightly.