SkillJumpState: simplify and match AgentSkillExecJumpState API
Removed the different instantiation types, only accept a skills list of
skills to execute and modify this list to take the very same parameters
as the AgentSkillExecJumpState. Implemented least commitment (again,
similar to luaagent). These changes make it easier to understand and
simpler to use. Changed the parameters final_state to final_to, and
failure_state to fail_to aligning it further.

This change is disruptive and will likely break some code, hence work on
topic branch for testing and integration.